3-26-0:27

+++ Accounting/views.py
@@ -15,6 +15,25 @@ class IndexView(generic.ListView):
 def DetailView(request,event_id):
     Event = get_object_or_404(event,pk = event_id)
     Members = Event.Member.all()
+    form = addPayForm(Event,request.POST)
+
+    if "delete" in request.POST:
+        Pay = Event.pay_set.get(pk = request.POST['delete'])
+        Pay.delete()
+
+    if form.is_valid():
+        try:
+            APay = Event.pay_set.create(
+                payer = form.cleaned_data["payer"],
+                price = int(form.cleaned_data["price"]),
+                purpose = form.cleaned_data["purpose"]
+                )
+            for i in form.cleaned_data["payee"]:
+                APay.payee.add(Members.get(name = i))
+            APay.save()
+        except:
+            pass
+
     pays = Event.pay_set.all().order_by("id")
     #value["A"]["B"]でAがBに払うべき金額を示すような二次元辞書を作る
     value = dict()
@@ -64,68 +83,26 @@ def DetailView(request,event_id):
     template_name = 'Accounting/detail.html'
     return render(request,template_name,context)
 
-def changeDB(request,event_id):
-    Event = get_object_or_404(event,pk=event_id)
-    Members = Event.Member.all()
-    form = addPayForm(Event,request.POST)
-    if form.is_valid():
-        try:
-            APay = Event.pay_set.create(
-                payer = form.cleaned_data["payer"],
-                price = int(form.cleaned_data["price"]),
-                purpose = form.cleaned_data["purpose"]
-                )
-            APay.save()
-        except:
-            return redirect("Accounting:detail",event_id = event_id)
-    for i in form.cleaned_data["payee"]:
-        APay.payee.add(Members.get(name = i))
-    return redirect("Accounting:detail",event_id = event_id)
 
-def deletePay(request,event_id):
+
+
+def history(request,event_id):
     Event = get_object_or_404(event,pk=event_id)
-    try:
+
+    if "delete" in request.POST:
         Pay = Event.pay_set.get(pk = request.POST['delete'])
         Pay.delete()
-    except:
-        return redirect("Accounting:detail",event_id = event_id)
-    return redirect("Accounting:detail",event_id = event_id)
 
-
-def deletePayHistory(request,event_id):
-    Event = get_object_or_404(event,pk=event_id)
-    try:
-        Pay = Event.pay_set.get(pk = request.POST['delete'])
-        Pay.delete()
-    except:
-        return redirect("Accounting:history",event_id = event_id)
-    return redirect("Accounting:history",event_id = event_id)
-
-
-def historySearch(request,event_id):
-    Event = get_object_or_404(event,pk=event_id)
+    pays = Event.pay_set.all().order_by("id")
     Menbers = Event.Member.all()
     form = searchHistoryForm(Event,request.POST)
-    pays = Event.pay_set.all().order_by("id")
+    template_name = 'Accounting/history.html'
     if form.is_valid():
         if form.cleaned_data["payer"]:
             pays = pays.filter(payer = Menbers.get(name = form.cleaned_data["payer"]))
         if form.cleaned_data["payee"]:
             for i in form.cleaned_data["payee"]:
                 pays = pays.filter(payee = Menbers.get(name = i))
-
-    template_name = 'Accounting/history.html'
-    context ={
-        "event":Event,
-        "pays":pays,
-        "form":searchHistoryForm(Event)
-    }
-    return render(request,template_name,context)
-
-def history(request,event_id):
-    Event = get_object_or_404(event,pk=event_id)
-    pays = Event.pay_set.all().order_by("id")
-    template_name = 'Accounting/history.html'
     context ={
         "event":Event,
         "pays":pays,
